Fixing GradientCheck
